# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "AgentGuard \u2014 Sub-10ms Agent Security Layer",
    "description": "Real-time enterprise policy and safety gateway for autonomous AI agents. Powered by Moss in-process sub-10ms hybrid semantic search.",
    "version": "1.0.0"
  },
  "paths": {
    "/api/v1/guard/evaluate": {
      "post": {
        "summary": "Evaluate an AI agent action against Moss security policies in <10ms",
        "operationId": "evaluate_action_api_v1_guard_evaluate_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ActionEvaluationRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ActionEvaluationResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/policies": {
      "get": {
        "summary": "List all security policies loaded in AgentGuard and indexed in Moss",
        "operationId": "list_policies_api_v1_policies_get",
        "parameters": [
          {
            "name": "category",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filter by policy category",
              "title": "Category"
            },
            "description": "Filter by policy category"
          },
          {
            "name": "enforcement",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filter by enforcement verdict",
              "title": "Enforcement"
            },
            "description": "Filter by enforcement verdict"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Policy"
                  },
                  "title": "Response List Policies Api V1 Policies Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Create and dynamically index a new policy into Moss runtime",
        "operationId": "create_policy_api_v1_policies_post",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PolicyCreateRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Policy"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/approvals/{approval_id}/decide": {
      "post": {
        "summary": "Human reviewer submits decision (APPROVE or REJECT) on a paused action",
        "operationId": "decide_approval_api_v1_approvals__approval_id__decide_post",
        "parameters": [
          {
            "name": "approval_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "description": "Approval UUID",
              "title": "Approval Id"
            },
            "description": "Approval UUID"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ApprovalDecisionRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApprovalDecisionResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/benchmark": {
      "post": {
        "summary": "Execute side-by-side latency test: In-Process Moss (<10ms) vs Remote Vector DB",
        "operationId": "run_latency_benchmark_api_v1_benchmark_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BenchmarkRequest",
                "default": {
                  "iterations": 50,
                  "warmup": 5
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BenchmarkResult"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/simulator/run": {
      "post": {
        "summary": "Run a simulation scenario and return evaluation result",
        "operationId": "run_scenario_api_v1_simulator_run_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ScenarioRunRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ActionEvaluationResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/interceptions": {
      "get": {
        "summary": "Fetch all real intercepted agent actions recorded in SQLite audit log",
        "operationId": "list_interceptions_api_v1_interceptions_get",
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 200,
              "minimum": 1,
              "default": 50,
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "ActionEvaluationRequest": {
        "properties": {
          "agent_id": {
            "type": "string",
            "title": "Agent Id",
            "example": "customer-support-bot"
          },
          "agent_role": {
            "type": "string",
            "title": "Agent Role",
            "example": "tier_1_support"
          },
          "tool_name": {
            "type": "string",
            "title": "Tool Name",
            "example": "stripe_issue_refund"
          },
          "parameters": {
            "additionalProperties": true,
            "type": "object",
            "title": "Parameters",
            "example": {
              "amount": 1500,
              "customer_id": "cust_88219"
            }
          },
          "context": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Context",
            "example": "Customer requested full refund due to order delay."
          },
          "dry_run": {
            "type": "boolean",
            "title": "Dry Run",
            "description": "If true, do not persist to audit log or trigger real approvals",
            "default": false
          }
        },
        "type": "object",
        "required": [
          "agent_id",
          "agent_role",
          "tool_name"
        ],
        "title": "ActionEvaluationRequest"
      },
      "ActionEvaluationResponse": {
        "properties": {
          "action_id": {
            "type": "string",
            "title": "Action Id"
          },
          "verdict": {
            "$ref": "#/components/schemas/GuardVerdict"
          },
          "risk_level": {
            "$ref": "#/components/schemas/RiskLevel"
          },
          "risk_score": {
            "type": "integer",
            "maximum": 100.0,
            "minimum": 0.0,
            "title": "Risk Score"
          },
          "reason": {
            "type": "string",
            "title": "Reason"
          },
          "matched_policies": {
            "items": {
              "$ref": "#/components/schemas/MatchedPolicy"
            },
            "type": "array",
            "title": "Matched Policies"
          },
          "approval_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Approval Id"
          },
          "execution_result": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ToolExecutionResult"
              },
              {
                "type": "null"
              }
            ]
          },
          "latency": {
            "$ref": "#/components/schemas/LatencyBreakdown"
          },
          "timestamp": {
            "type": "string",
            "title": "Timestamp"
          }
        },
        "type": "object",
        "required": [
          "action_id",
          "verdict",
          "risk_level",
          "risk_score",
          "reason",
          "matched_policies",
          "latency",
          "timestamp"
        ],
        "title": "ActionEvaluationResponse"
      },
...
      "ApprovalDecisionRequest": {
        "properties": {
          "decision": {
            "type": "string",
            "pattern": "^(APPROVE|REJECT)$",
            "title": "Decision"
          },
          "reviewer_notes": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reviewer Notes"
          },
          "reviewed_by": {
            "type": "string",
            "title": "Reviewed By",
            "example": "admin@agentguard.dev"
          }
        },
        "type": "object",
        "required": [
          "decision",
          "reviewed_by"
        ],
        "title": "ApprovalDecisionRequest"
      },
      "ApprovalDecisionResponse": {
        "properties": {
          "approval_id": {
            "type": "string",
            "title": "Approval Id"
          },
          "action_id": {
            "type": "string",
            "title": "Action Id"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "resolved_by": {
            "type": "string",
            "title": "Resolved By"
          },
          "resolved_at": {
            "type": "string",
            "title": "Resolved At"
          },
          "reviewer_notes": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reviewer Notes"
          },
          "moss_memory_updated": {
            "type": "boolean",
            "title": "Moss Memory Updated",
            "default": true
          },
          "execution_result": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ToolExecutionResult"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "type": "object",
        "required": [
          "approval_id",
          "action_id",
          "status",
          "resolved_by",
          "resolved_at"
        ],
        "title": "ApprovalDecisionResponse"
      },
      "BenchmarkRequest": {
        "properties": {
          "iterations": {
            "type": "integer",
            "title": "Iterations",
            "default": 50
          },
          "warmup": {
            "type": "integer",
            "title": "Warmup",
            "default": 5
          }
        },
        "type": "object",
        "title": "BenchmarkRequest"
      },
      "BenchmarkResult": {
        "properties": {
          "total_queries": {
            "type": "integer",
            "title": "Total Queries"
          },
          "moss": {
            "$ref": "#/components/schemas/LatencyStats"
          },
          "remote_vector_db": {
            "$ref": "#/components/schemas/LatencyStats"
          },
          "speedup_factor": {
            "type": "number",
            "title": "Speedup Factor"
          },
          "timestamp": {
            "type": "string",
            "title": "Timestamp"
          }
        },
        "type": "object",
        "required": [
          "total_queries",
          "moss",
          "remote_vector_db",
          "speedup_factor",
          "timestamp"
        ],
        "title": "BenchmarkResult"
      },
      "GuardVerdict": {
        "type": "string",
        "enum": [
          "ALLOW",
          "BLOCK",
          "REQUIRE_APPROVAL"
        ],
        "title": "GuardVerdict"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "Policy": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "category": {
            "$ref": "#/components/schemas/PolicyCategory"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "rule_text": {
            "type": "string",
            "title": "Rule Text"
          },
          "enforcement": {
            "$ref": "#/components/schemas/GuardVerdict"
          },
          "risk_level": {
            "$ref": "#/components/schemas/RiskLevel"
          },
          "target_tools": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Target Tools"
          },
          "conditions": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Conditions"
          },
          "is_active": {
            "type": "boolean",
            "title": "Is Active",
            "default": true
          },
          "created_at": {
            "type": "string",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "category",
          "name",
          "rule_text",
          "enforcement",
          "risk_level",
          "target_tools",
          "created_at"
        ],
        "title": "Policy"
      },
...
      "PolicyCreateRequest": {
        "properties": {
          "category": {
            "$ref": "#/components/schemas/PolicyCategory"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "rule_text": {
            "type": "string",
            "title": "Rule Text"
          },
          "enforcement": {
            "$ref": "#/components/schemas/GuardVerdict",
            "default": "BLOCK"
          },
          "risk_level": {
            "$ref": "#/components/schemas/RiskLevel",
            "default": "HIGH"
          },
          "target_tools": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Target Tools"
          },
          "conditions": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Conditions"
          },
          "is_active": {
            "type": "boolean",
            "title": "Is Active",
            "default": true
          }
        },
        "type": "object",
        "required": [
          "category",
          "name",
          "rule_text"
        ],
        "title": "PolicyCreateRequest"
      },
...
      "ScenarioRunRequest": {
        "properties": {
          "scenario_id": {
            "type": "string",
            "title": "Scenario Id"
          }
        },
        "type": "object",
        "required": [
          "scenario_id"
        ],
        "title": "ScenarioRunRequest"
      },
...
    }
  }
}